pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-smyth-technical"
version = "0.1.0"
description = "A RAG Agent API for processing tweets and answering queries"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "UNLICENCED"}
authors = [
    {name = "David Alexander", email = "dcalex30@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.2",
    "langchain>=0.0.267",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "openai>=1.3.0",
    "faiss-cpu>=1.7.4",
    "tiktoken>=0.5.0",
    "aws-cdk-lib>=2.85.0",
    "constructs>=10.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
]
infra = [
    "aws-cdk-lib>=2.85.0",
    "constructs>=10.2.0",
    "pytest-cdk>=0.0.1",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[project.scripts]
agent-api = "src.main:run_app"

[tool.hatch.build.targets.wheel]
packages = ["src"]
